requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
identify==2.2.9
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==20.9
pathspec==0.8.1
pre-commit==2.13.0
psycopg2-binary==2.8.6
pycparser==2.20
pygraphviz==1.7
pyjsend==0.2.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
python-jsend-response==0.0.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
